Add support for initial Doxygen documentation for c_salt.
This is meant to be a starting point for allowing developers to easily generate and view Doxygen documentation from the c_salt example.  It is not meant to be release-quality documentation.

This code is copied from http://src.chromium.org/svn/trunk/src/ppapi/documentation r67913, with modifications to change ppapi to c_salt.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/5876003