Changed SDK to now build the ppapi and nacl library files.
Previously, the SDK was relying on the version of the libraries as bundled
in the toolchain.  However, the ppapi header files are out of sync with these
libraries, so it is better to rebuild them with the correct headers.

Also, this change migrates the building of sel_ldr and ncval into
./scons toolchain.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=1861
TEST= Trybots

Review URL: http://codereview.chromium.org/7198036