Add ability to add Nmf and Javascript files to a project. Rename NexePlugin to WebApp. Set DebugHost  (chrome for a WebApp, sel_ldr for a ConsoleAPp). Include an NMF file in a WebApp by default.
Also some directory cleanup -- the console application had an extra layer of file directory that did match the VS project organization and made it hard to view/edit the console files.

Note the .cc and .js files are just empty or a copy of stub -- to get the user started.

The *real* source files here are the .naclproj files and .vstemplate files, which indicate which files are included in a given project (in the Templates/Projects directory) or for the Templates/ProjectItems directory point to the sample file.

Follow up CLs will work on:
  1) Smart renaming of files / contents based on Project name
  2) For web-based project, deploying html/nexe/nmf files to a common dir before launching server and chrome.
Review URL: http://codereview.chromium.org/7464017