Template changes for generating names and contents based on Project Name.
BUG= https://tracker.corp.google.com/story/show/391458
TEST= manual, since this gets called by VSX to generate projects

The .c and .cc files are copied from the project_templates directory in the SDK and then modified to use $safeprojectname$ in the appropriate places.

The .vstemplate files use $safeprojectname$ to replace the basic files in the Project template with a filename based on the project name.

The .html file and .nmf files also use $safeprojectname$.
Review URL: http://codereview.chromium.org/7491025