Add a function to get the table of contents for a tar archive.
This is the first of a few CLs to support symlinks and hardlinks in the NSIS
installer.  Follow-on CLs will preserve the tar archives downloaded by
download_toolchains.py, and then use this utility to get the list of files,
hard links and symlinks from them.  Once that happens, I can update the NSIS
installer to use this info to produce correct symlinks (and hardlinks) in the
installer.
Review URL: http://codereview.chromium.org/7792079