Added Roland's generate_manifest.py to the SDK
Note:  This version only checks in Roland's original
generate_manifest.py (renamed to create_nmf.py) with
minimal changes and a simple unit test.  It does not tie
into the SCons code yet.  Those changes are planned for
future check-ins.

I felt it would be useful to commit this as a baseline,
and then iterate on top of this.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2299
BUG= http://code.google.com/p/nativeclient-sdk/issues/detail?id=137
TEST=nmf_test.py

Review URL: http://codereview.chromium.org/8229034