There are two things in this review: Unit Tests to verify that project settings default to the correct values, and a patching system to copy Win32 settings to create the PPAPI template.
Note the only code that needs reviewing is the .py and .cs files.

The other files are patch xml files that create the PPAPI template.

Added the patching system to copy the Win32 settings to create the PPAPI settings.

Removed the vsmdi file since it is better as a developer dependent file.

Added unit tests to verify the default settings for PPAPI and NaCl platforms.

Added ProjectSettingsTest project

BUG=136414
TEST=

Review URL: https://chromiumcodereview.appspot.com/10797040