Fix for Visual Studio prompting for $(CHROME_PATH).
We need to make sure that the expanded version of
$(CHROME_PATH) is always used, otherwise Visual Studio will
complain, so we expand this variable on within the
"LocalDebuggerCommand" as part of the project modifications
that we do when NaCl/PPAPI are first added to the project.

The problem is that the "LocalDebuggerCommand" is stored not
in the project itself but in the .user file.  This change
means that even if the user deletes that .user file (or more
likely receives a project files without an accompanying
.user file) we detect that $(CHROME_PATH) is within
"LocalDebuggerCommand" and force an update to the settings
at that time.

BUG=155094

Review URL: https://codereview.chromium.org/11192036